# cmd/controller/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Qryptic Controller Service to manage the users, groups, gateways and clients.",
        "title": "Qryptic Controller API",
        "contact": {
            "name": "Leetsecure Support",
            "url": "leetsecure.com",
            "email": "hello@leetsecure.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/group": {
            "post": {
                "description": "CreateGroup",
                "tags": [
                    "admin-group"
                ],
                "summary": "CreateGroup",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Group details",
                        "name": "GroupCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/group/list": {
            "get": {
                "description": "list all the groups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-group"
                ],
                "summary": "ListGroups",
                "operationId": "ListGroups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "string collections",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/group/{id}": {
            "get": {
                "description": "Get Group By UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-group"
                ],
                "summary": "Get Group",
                "operationId": "GetGroup",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "string collections",
                        "name": "include",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "UpdateGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-group"
                ],
                "summary": "UpdateGroup",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Group details",
                        "name": "GroupUpdateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-group"
                ],
                "summary": "DeleteGroup",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/group/{id}/{action}/users": {
            "put": {
                "description": "AddRemoveUsersInGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-group"
                ],
                "summary": "AddRemoveUsersInGroup",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Provide details",
                        "name": "GroupUpdateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupUpdateUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "action",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Client": {
            "type": "object",
            "properties": {
                "allocatedIP": {
                    "type": "string"
                },
                "allowedIPs": {
                    "type": "string"
                },
                "clientPrivateKey": {
                    "type": "string"
                },
                "clientPublicKey": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dnsServer": {
                    "type": "string"
                },
                "expiryTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "preshared_key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.User"
                },
                "userId": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                },
                "vpnGateway": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGateway"
                },
                "vpnGatewayId": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Group": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.User"
                    }
                },
                "uuid": {
                    "type": "string"
                },
                "vpnGateways": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGateway"
                    }
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.GroupUpdateUserRequest": {
            "type": "object",
            "required": [
                "userUuids"
            ],
            "properties": {
                "userUuids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.IPPool": {
            "type": "object",
            "properties": {
                "assigned": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "vpngatewayID": {
                    "description": "Foreign key to VpnGateway",
                    "type": "integer"
                }
            }
        },
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.User": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Client"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Group"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "isPasswordSet": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UserRoleEnum"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "vpnGateways": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGateway"
                    }
                }
            }
        },
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UserRoleEnum": {
            "type": "string",
            "enum": [
                "Default",
                "Admin",
                "User"
            ],
            "x-enum-varnames": [
                "DefaultRole",
                "AdminRole",
                "UserRole"
            ]
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGateway": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Client"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dnsServer": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.Group"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "ipAddressCIDR": {
                    "type": "string"
                },
                "ipPool": {
                    "description": "IPAllocations    []*IPAllocation `json:\"ipAllocations\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.IPPool"
                    }
                },
                "jwtAlgorithm": {
                    "type": "string"
                },
                "jwtSecretKey": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "serverPrivateKey": {
                    "type": "string"
                },
                "serverPublicKey": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.User"
                    }
                },
                "uuid": {
                    "type": "string"
                },
                "vpnCIDR": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}